% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_options.R
\name{get_opts}
\alias{get_opts}
\title{Get Options}
\usage{
get_opts(opts = NULL)
}
\arguments{
\item{opts}{A character vector of option names to get. Default is \code{NULL},
and if left \code{NULL}, will return all available package options included
in \code{noclocksrtools}.}
}
\value{
A named list of package options
}
\description{
Get package options
}
\examples{
get_opts()
}
