% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_sys.R
\name{pkg_sys_ext}
\alias{pkg_sys_ext}
\title{Package System Extensions}
\description{
Package System Extensions
}
\section{Extensions}{


\code{noclocksrtools} also includes a variety of convenience \emph{extension function}
which wrap \code{pkg_sys()} to access specific directories or types of package
resources.
\subsection{External Data}{
\itemize{
\item \code{pkg_sys_extdata()}: Access package data files in the \verb{extdata/} external data folder.
}
}

\subsection{Templates}{
\itemize{
\item \code{pkg_sys_templates()}: Access package templates in the \code{inst/templates} folder.
}
}

\subsection{Scripts}{
\itemize{
\item \code{pkg_sys_scripts()}: Access package scripts in the \code{inst/scripts} folder.
}
}

\subsection{Configuration}{
\itemize{
\item \code{pkg_sys_config()}: Access package configuration files in the \code{inst/config} folder.
}
}

\subsection{Documentation}{
\itemize{
\item \code{pkg_sys_docs()}: Access package documentation files in the \code{inst/docs} folder.
}

There is also the functions \code{\link[=pkg_docs]{pkg_docs()}} and \code{\link[=pkg_manual]{pkg_manual()}} which launch
the package's locally installed documentation website (via \code{pkgdown}) and
its manual PDF, respectively.
}

\subsection{Examples}{
\itemize{
\item \code{pkg_sys_examples()}: Access package example files in the \code{inst/examples} folder.
}
}

\subsection{Assets}{
\itemize{
\item \code{pkg_sys_assets()}: Access package assets included in the \verb{assets/} folder.
}
}
}

