% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_sys.R
\name{system_file}
\alias{system_file}
\title{System File}
\usage{
system_file(...)
}
\description{
This function is actually a \code{\link[purrr:partial]{purrr::partial()}} which re-implements
\code{\link[base:system.file]{base::system.file()}} to be pre-configured to access files from this
package: \code{noclocksrtools}.
}
\section{Implementation}{

This function is implemented using \code{\link[purrr:partial]{purrr::partial()}} to create a partial
function which wraps \code{\link[base:system.file]{base::system.file()}} and sets the \code{package} argument
to \code{noclocksrtools}.
}

\seealso{
\itemize{
\item \code{\link[=pkg_sys]{pkg_sys()}}
\item \code{\link[purrr:partial]{purrr::partial()}}
\item \code{\link[base:system.file]{base::system.file()}}, \code{\link[fs:path_package]{fs::path_package()}}
}
}
\concept{Package Utilities}
\keyword{internal}
